refactor(config): extract GoogleChat schema into zod-schema.providers-googlechat.ts
Move GoogleChat channel schemas (GoogleChatDmSchema, GoogleChatGroupSchema,
GoogleChatAccountSchema, GoogleChatConfigSchema) from the monolithic
zod-schema.providers-core.ts into a dedicated zod-schema.providers-googlechat.ts
module, following the same pattern as zod-schema.providers-whatsapp.ts.

Also exports BotLoopProtectionSchema from providers-core since GoogleChat
shares it with other channel schemas.
