+++ src/plugin-sdk/bundled-channel-config-schema.ts
@@ -25,11 +25,11 @@ export {
 export { ToolPolicySchema } from "../config/zod-schema.agent-runtime.js";
 export {
   DiscordConfigSchema,
-  GoogleChatConfigSchema,
   IMessageConfigSchema,
   MSTeamsConfigSchema,
   SignalConfigSchema,
   SlackConfigSchema,
   TelegramConfigSchema,
 } from "../config/zod-schema.providers-core.js";
+export { GoogleChatConfigSchema } from "../config/zod-schema.providers-googlechat.js";
 export { WhatsAppConfigSchema } from "../config/zod-schema.providers-whatsapp.js";
